pyproject.toml
[project]
name = "mcp-a2a-udemy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.12",
    "fastmcp>=2.7.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.25",
    "langchain-core>=0.3.60",
    "langchain-mcp-adapters>=0.1.1",
    "langchain-openai>=0.3.17",
    "langgraph>=0.4.5",
    "openai-agents>=0.2.3",
    "python-jose>=3.4.0",
    "ruff>=0.11.10",
    "uvicorn>=0.34.2",
]
